Created an activity prototype. Renamed classes.
