+++ app/src/main/java/com/example/android/app/AttributeScreenActivity.java
@@ -2,6 +2,12 @@ package com.example.android.app;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.widget.TextView;
+
+import com.example.model.character.Character;
+import com.example.model.character.CharacterStatus;
+import com.example.model.character.status.energy.Energy;
+import com.example.model.character.status.health.Health;
 
 public class AttributeScreenActivity extends AppCompatActivity {
