--- app/src/main/java/com/example/model/character/attribute/resilence/Resilence.java
@@ -1,14 +1,0 @@
-package com.example.model.character.attribute.resilence;
-
-import com.example.model.character.attribute.Attribute;
-
-/**
- * Class representing the Resilence attribute.
- *
- * @author Lucas Andrade
- */
-public class Resilence extends Attribute {
-    public Resilence(Long initialValue) {
-        super(initialValue);
-    }
-}
