+++ app/src/main/java/com/example/model/character/attribute/resilience/Resilience.java
@@ -1,0 +1,14 @@
+package com.example.model.character.attribute.resilience;
+
+import com.example.model.character.attribute.Attribute;
+
+/**
+ * Class representing the Resilience attribute.
+ *
+ * @author Lucas Andrade
+ */
+public class Resilience extends Attribute {
+    public Resilience(Long initialValue) {
+        super(initialValue);
+    }
+}
